Renamed utils and discovery packages to internal in the gardena binding. See openhab/static-code-analysis#100
